Add files via upload
This is the first version of this project. So far, a user can be created and a product can be added to the page using an admin account which must be created. To use these files a virtual environment must be set up and activated. More information on how to do this is given in the VirtualEnvAndDjango.pdf file. The VirtualEnvAndDjango.pdf is a set of notes that I've created to detail the process of working with virtual envs and django. These notes are written in such a way that they should make sense to anyone that reads them but they are primarily written for myself so the tone is a bit less formal but the explanations should still be clear enough to follow along with.